spotify-aura-backend/requirements.txt
fastapi
uvicorn
spotipy
python-dotenv
opencv-python
mediapipe
pydantic
python-multipart